refactor(release-scoring): move the scoring kernel into common
`modules/media/release-rejection.helper.ts` was already shared, not
media-specific: `profiles.service.ts` imports `buildAllowedQualityIds` and
`allowedAudioLanguageIds` from it, and `media-query.service.ts` imports
`rankFromQualityString`. A leaf module reaching back into `media` for its
own domain logic is the wrong direction, and it made `media` impossible to
reason about as a boundary.

Two edges are inverted rather than carried along. `TorznabRelease` is
declared in `indexers/torznab.service.ts` yet describes the contract between
whatever fetches releases and whatever scores them; it moves into the kernel
as `ReleaseCandidate`, and the fetcher now imports it. `buildIndexerMinSeeders`
took the `Indexer` entity when it reads only an id and one setting, so it now
takes that shape structurally — every caller still passes its entities with
no cast.

Pure move plus renames; no behaviour change. The kernel no longer imports
anything from the indexer or download-client modules. Its two remaining
cross-module imports are quality and language profile item types, which are
core.
